agent/verifier-agent/requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.8
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
together==1.4.0
uvicorn==0.30.1